update pyo3 macros
