+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
-  "python.pythonPath": "/Users/liuchang/mambaforge/envs/crystallus/bin/python",
-  "cSpell.words": [
-    "consts"
-  ]
+    "python.pythonPath": "/Users/liuchang/mambaforge/envs/crystallus/bin/python",
+    "cSpell.words": [
+        "consts",
+        "pyfn"
+    ]
 }
