+++ Cargo.toml
@@ -38,7 +38,7 @@ crate-type = ['cdylib']
 name = 'crystallus'
 
 [dependencies]
-itertools = '0.10.0'
+itertools = '0.10.3'
 rayon = '1.5.1'
 
 [dependencies.pyo3]
